Add toolbar with a colored Activity button and a working Stop button
This change will work only if the libraries sugar-html-core and
sugar-html-graphics are added (with volo add).  We can consider adding
them directly to this repo, as git submodules.

The js code to colorize the activity icon may be too large.  We could
provide a better API.

Ideally we should define the SVG graphics in the CSS.  Looks doable:
Daniel Narvaez did it in Agora.  But I couldn't in my pet activity,
seems that the content inside the the "data:" uri scheme is bad.  Here
is the branch in which I was trying:

https://github.com/manuq/clockjs/tree/colorize-css